Somatic and HC GenotypingEngine only share methods that need to be shared
